+++ src/main/java/racingcar/Application.java
@@ -1,12 +1,20 @@
 package racingcar;
 
-import racingcar.controller.RaceController;
+import racingcar.controller.Race;
+import racingcar.domain.Cars;
+import racingcar.domain.MoveCount;
+import racingcar.view.InputView;
+import racingcar.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
-        RaceController raceController = new RaceController();
-        raceController.initRace();
-        raceController.run();
+        OutputView.printRacingStart();
+        Cars cars = new Cars(InputView.getCarNames());
+
+        OutputView.printAskCount();
+        MoveCount moveCnt = InputView.inputMoveCount();
+
+        Race race = new Race(cars, moveCnt);
+        race.run();
     }
 }
-
